# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookLanguage = typeof BookLanguage[keyof typeof BookLanguage];


export const BookLanguage = {
  arabic: 'arabic',
  english: 'english',
} as const;

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  arabicTitle: string | null;
  author: string;
  language: BookLanguage;
  category: string;
  sourcePrice: number;
  sellingPrice: number;
  sourceUrl: string;
  /** @nullable */
  description: string | null;
  featured: boolean;
  coverTone: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CatalogSummary {
  total: number;
  arabic: number;
  english: number;
  featured: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  sourcing: 'sourcing',
  ready: 'ready',
  handed: 'handed',
} as const;

export interface Order {
  id: number;
  bookId: number;
  bookTitle: string;
  studentName: string;
  className: string;
  contact: string;
  quantity: number;
  price: number;
  status: OrderStatus;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface OrderInput {
  bookId: number;
  /** @minLength 2 */
  studentName: string;
  /** @minLength 1 */
  className: string;
  /** @minLength 3 */
  contact: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  sourcing: 'sourcing',
  ready: 'ready',
  handed: 'handed',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface OrderSummary {
  total: number;
  pending: number;
  sourcing: number;
  ready: number;
  handed: number;
}

export type ListBooksParams = {
/**
 * Search by English or Arabic title, author, or category
 */
search?: string;
/**
 * Filter by catalog language
 */
language?: ListBooksLanguage;
/**
 * Filter by category
 */
category?: string;
/**
 * Only return featured books
 */
featured?: boolean;
/**
 * Results page
 * @minimum 1
 */
page?: number;
/**
 * Results per page
 * @minimum 1
 * @maximum 120
 */
limit?: number;
};

export type ListBooksLanguage = typeof ListBooksLanguage[keyof typeof ListBooksLanguage];


export const ListBooksLanguage = {
  all: 'all',
  arabic: 'arabic',
  english: 'english',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  all: 'all',
  pending: 'pending',
  sourcing: 'sourcing',
  ready: 'ready',
  handed: 'handed',
} as const;