# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookLanguage } from './bookLanguage';

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  arabicTitle: string | null;
  author: string;
  language: BookLanguage;
  category: string;
  sourcePrice: number;
  sellingPrice: number;
  sourceUrl: string;
  /** @nullable */
  description: string | null;
  featured: boolean;
  coverTone: string;
  /** @nullable */
  imageUrl?: string | null;
}