# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogSummary {
  total: number;
  arabic: number;
  english: number;
  featured: number;
}