# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  bookId: number;
  bookTitle: string;
  studentName: string;
  className: string;
  contact: string;
  quantity: number;
  price: number;
  status: OrderStatus;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
}