# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  bookId: number;
  /** @minLength 2 */
  studentName: string;
  /** @minLength 1 */
  className: string;
  /** @minLength 3 */
  contact: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  quantity: number;
  /** @nullable */
  notes?: string | null;
}