# lib/api-zod/src/generated/types/orderSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderSummary {
  total: number;
  pending: number;
  sourcing: number;
  ready: number;
  handed: number;
}